LayoutTests/http/tests/dom/opened-window-not-targetable-after-disowning-opener-expected.txt

Make sure that windows opened via window.open are not targetable by their opener after it is disowned. If testing manually, you should see 2 tabs open.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS w.location.href is "about:blank"
PASS testRunner.windowCount() is 3
PASS successfullyParsed is true

TEST COMPLETE
